+++ src/pages/admin/modals/UserModal.jsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState, useRef } from "react";
-import Button from "../../components/Button"; 
+import Button from "../../../components/Button"; 
 
 const UserModal = ({ open, onClose, onSubmit, initialData, readOnly }) => {
   const modalRef = useRef(null);
