+++ src/router/Router.jsx
@@ -9,8 +9,8 @@ import { authGuard, adminGuard } from "../validators/routeValidator";
 import SubscriptionCheckout from "../pages/SubscriptionCheckout";
 import MainPage from "../pages/MainPage";
 import AdminDashboard from "../pages/admin/AdminDashboard";
-import UsersTab from "../pages/admin/UsersTab";
-import PlansTab from "../pages/admin/PlansTab";
+import UsersTab from "../pages/admin/tables/UsersTab";
+import PlansTab from "../pages/admin/tables/PlansTab";
 import OrdersTab from "../pages/admin/OrdersTab";
 import PaymentsTab from "../pages/admin/PaymentsTab";
 import GiftPage from "../pages/GiftPage";
